requirements.txt
django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.11
python-telegram-bot==13.15
sentry-sdk==1.9.7
